# Optimize JFR Type Repository Serialization