--- a/substratevm/src/com.oracle.svm.core/src/com/oracle/svm/core/hub/PredefinedClassesSupport.java
+++ b/substratevm/src/com.oracle.svm.core/src/com/oracle/svm/core/hub/PredefinedClassesSupport.java
@@ -102,6 +102,8 @@ public final class PredefinedClassesSupport {
     /** Predefined classes which have already been loaded, by name. */
     private final EconomicMap<String, Class<?>> loadedClassesByName = EconomicMap.create();
 
+    private long loadedClassesCount;
+
     @Platforms(Platform.HOSTED_ONLY.class)
     public static void registerClass(String hash, Class<?> clazz) {
         Class<?> existing = singleton().predefinedClassesByHash.putIfAbsent(hash, clazz);
@@ -123,6 +125,10 @@ public final class PredefinedClassesSupport {
     @Platforms(Platform.HOSTED_ONLY.class)
     public static boolean isPredefined(Class<?> clazz) {
         return singleton().predefinedClasses.contains(clazz);
+    }
+
+    public static long getRuntimeLoadedClassesCount() {
+        return singleton().loadedClassesCount;
     }
 
     public static Class<?> loadClass(ClassLoader classLoader, String expectedName, byte[] data, int offset, int length, ProtectionDomain protectionDomain) {
@@ -190,6 +196,7 @@ public final class PredefinedClassesSupport {
                 hub.setProtectionDomainAtRuntime(protectionDomain);
             }
             loadedClassesByName.put(clazz.getName(), clazz);
+            loadedClassesCount++;
             return true;
         } finally {
             lock.unlock();