@christianhaeubl 2023-08-31
Thanks for doing the measurements and for the PR. The list of loaded classes can be pretty big, so I would like to avoid having one extra copy in the heap. I replaced the method with a visitor, see https://github.com/oracle/graal/pull/7317